Convert project submission page to TSX
